agents.py
import os
from dotenv import load_dotenv
from crewai import Agent, LLM
from crewai.tools import tool
from tavily import TavilyClient

# 1. Load the keys from your .env file
load_dotenv()

# 2. Setup the Gemini "Brain" using CrewAI's native LLM class
gemini_llm = LLM(
    model="gemini/gemini-2.0-flash-lite",
    temperature=0.5,
    api_key=os.getenv("GEMINI_API_KEY")
)

# 3. Setup the Search Tool using CrewAI's native @tool decorator
@tool("Tavily Search")
def search_tool(query: str) -> str:
    """Search the web for current, accurate information. Input should be a search query string."""
    client = TavilyClient(api_key=os.getenv("TAVILY_API_KEY"))
    results = client.search(query=query, max_results=5)
    return str(results)

# ── ORIGINAL AGENTS ──────────────────────────────────────────────────────────

researcher = Agent(
    role='Market Research Analyst',
    goal=(
        'Find real-world 2026 competitors for {product_idea}. '
        'Go beyond surface-level names — identify pricing tiers, key differentiators, '
        'and specific gaps in the market that {product_idea} could exploit.'
    ),
    backstory=(
        'You are a senior analyst at a top-tier strategy consulting firm. '
        'You specialize in competitive intelligence and market sizing. '
        'You never report vague findings — every claim is backed by a source.'
    ),
    tools=[search_tool],
    llm=gemini_llm,
    verbose=True
)

tech_architect = Agent(
    role='ORIE Technical Architect',
    goal=(
        'Identify the 3 most critical technical bottlenecks and resource constraints '
        'for building {product_idea} at scale. '
        'Apply operations research and systems thinking — quantify where possible.'
    ),
    backstory=(
        'You hold a PhD in Operations Research and have architected systems at '
        'both early-stage startups and Fortune 500 companies. '
        'You think in queuing theory, optimization functions, and failure modes.'
    ),
    llm=gemini_llm,
    verbose=True
)

writer = Agent(
    role='Lead Product Manager',
    goal=(
        'Synthesize all research, financial analysis, UX findings, technical constraints, '
        'and risk flags into a single professional Markdown PRD for {product_idea}. '
        'Every section must be specific, actionable, and grounded in the prior agents\' outputs.'
    ),
    backstory=(
        'You are a Senior PM who has shipped products at Google, Stripe, and two YC startups. '
        'You write PRDs that engineers love and executives fund. '
        'You never write vague requirements — everything maps to a measurable outcome.'
    ),
    llm=gemini_llm,
    verbose=True
)

# ── NEW SPECIALIZED AGENTS ───────────────────────────────────────────────────

ux_researcher = Agent(
    role='UX Research Lead',
    goal=(
        'Define 3 detailed user personas for {product_idea} and map the primary user journey '
        'for each, identifying their core pain points, motivations, and where they currently fail '
        'with existing solutions. Use real behavioral data where possible.'
    ),
    backstory=(
        'You are a UX researcher with 10 years of experience running discovery sprints '
        'at IDEO and as Head of Research at a Series B SaaS company. '
        'You interview users, synthesize patterns, and translate human behavior into '
        'product requirements that resonate. You ground personas in real demographics, '
        'not stereotypes.'
    ),
    tools=[search_tool],
    llm=gemini_llm,
    verbose=True
)

financial_analyst = Agent(
    role='Startup Financial Analyst',
    goal=(
        'Produce a financial feasibility profile for {product_idea}: estimate the TAM/SAM/SOM, '
        'recommend a pricing model with 2-3 tiers, project a realistic MVP build cost range, '
        'and identify the key metrics needed to reach break-even. '
        'Use real market comparables found via search.'
    ),
    backstory=(
        'You are a former VC analyst turned startup CFO. You have evaluated over 300 pitch decks '
        'and built financial models for 12 funded companies. '
        'You are ruthlessly realistic — you call out vanity metrics and unrealistic TAM claims. '
        'You always cite comparable companies and funding benchmarks to ground your estimates.'
    ),
    tools=[search_tool],
    llm=gemini_llm,
    verbose=True
)

risk_analyst = Agent(
    role='Risk & Compliance Officer',
    goal=(
        'Identify the top 5 risks for {product_idea} across three categories: '
        'regulatory/legal (e.g. GDPR, HIPAA, FTC, SEC depending on domain), '
        'technical/security (data breaches, infrastructure failure, API dependencies), '
        'and market/execution (timing risk, adoption risk, competitor response). '
        'Assign each risk a severity (High/Medium/Low) and suggest a mitigation strategy.'
    ),
    backstory=(
        'You are a former cybersecurity attorney turned Chief Risk Officer. '
        'You have navigated GDPR audits, SOC 2 certifications, and FTC investigations. '
        'You think in threat models and probability-weighted impact. '
        'You are not alarmist — you only flag risks that are realistic and material.'
    ),
    tools=[search_tool],
    llm=gemini_llm,
    verbose=True
)

critic = Agent(
    role='Product Critic & Red Team Lead',
    goal=(
        'Review all prior research, UX findings, financial projections, technical analysis, '
        'and risk flags for {product_idea}. '
        'Challenge the 3 weakest assumptions, identify any blind spots or contradictions '
        'between agents\' outputs, and produce a concise red team report. '
        'Your job is to make the final PRD stronger — not to be contrarian for its own sake.'
    ),
    backstory=(
        'You are a seasoned product strategist who has killed more bad ideas than you have shipped. '
        'You were the "designated skeptic" on product reviews at Amazon and Netflix. '
        'You have a talent for spotting the assumption everyone else missed — '
        'the one that, if wrong, collapses the entire business case.'
    ),
    llm=gemini_llm,
    verbose=True
)